Improved heatmap2 function.
While explaining the function in the master thesis, I found and implemented some significant performance improvements. The function no-more uses joins, but arrays for preperating the user_pois table. Probably not needed at all, since the function is deprecated, but why not. Walkability is now calculated for all amenities in 900ms.